Created a doxygen mainpage and explaination of eci_interface.h.
Close #27
As an effort to begin comprehensive documentation of the ECI,
this commit creates a main page for doxygen and an explaination of how to build
eci_interface.h. Since the simple app's eci_interface.h was using normal
C style comments instead of doxygen comments, these were converted for
the documentation. So far, there wasn't a need to convert the sgp4 app
since it's not referenced. The last change was to remove the value after
every macro defined in the simple app's interface file. The comments mention
it can be anything but it makes more sense to just define the macro
without a value.