Fixed the submodule issue in the cmake build
When cloning the cFS repository we were doing a recusive clone
but this would get the submodules that were linked in master
and not our pinned commit. Now, the process is to clone cFS,
checkout out our commit, then recursively init the submodules.
Doing things in this order garentees we don't get the submodules
of master and we get the proper versions of our apps/tools.