Changed the behavior of non-ECI_CMD_MID messages
This change aims to remove the failure condition on messages
with a value command message id (0x1xxx). The limitations
that ECI was putting on messages registered with the MsgRcv
array doesn't allow for multiple commands and this change will
add the ability to create any number of commands with their
appropriate command structure.

The previous implmenation of the ECI commanding basically did
one check to see wether or not the command had a function
code between ECI_FUNC_CODE_START and ECI_FUNC_CODE_END but
then passed the message into the receive message function.
By directly passing the message with a command function code
into the rcv_msg function, the only check we lose is if that
command has a function code within the range. All other checks
for length and queue buffer size are still tested within this
function.

This commit also removes the test that enforces that expects an
error to be sent when receiving a command message that isn't
ECI_CMD_MID.